requirements.txt
flask==2.0.1
torch==1.9.0
torchvision==0.10.0
pillow==8.3.1
numpy==1.21.1
pyttsx3==2.90
pandas>=1.3.0
Pillow>=8.3.0
scikit-learn>=0.24.0
seaborn>=0.11.0
matplotlib>=3.4.0 